Backend/requirements.txt
fastapi==0.72.0
uvicorn==0.17.0
python-multipart==0.0.5
six==1.16.0
requests==2.27.1
pyyaml==6.0
pydantic==1.9.0
starlette==0.17.1
sqlalchemy==1.4.31
sqlalchemy-utils==0.38.2
alembic==1.7.5
addict==2.4.0
cryptography==36.0.1
python-jose==3.3.0
passlib==1.7.4
emails==0.6
pydantic[email]==1.9.0
Jinja2==3.0.3
boto3==1.21.20
psycopg2-binary==2.8.6
mangum==0.14.1
bcrypt==3.2.0
botocore~=1.24.24
authlib~=1.0.1
httpx==0.22.0
itsdangerous==2.1.2